requirements.txt
streamlit
pandas
openai
chromadb==0.4.24
scikit-learn
pysqlite3-binary; sys_platform == "linux"
protobuf==4.25.3
grpcio==1.62.1
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
python-dotenv
numpy<2.0.0
